Update the 'expandurl' function to use longurl.org
I've decided that I want to remain anonymous to all of the shortened URL
services. So, by using longurl.org, the shortened URL services will not be
able to track my IP address or other metrics. Of course, once longurl.org
is no longer online, this function will break, and I'll need to update it.
Of course, I'm not anonymous to longurl.org.