Replace ASCII coins with full-width Unicode
Now that I have the "woozy" emoji face to represent the drunk bishop, the prompt text is pushed to
the right to accommodate the width of the emoji character. A fix is to replace the ASCII with
full-width Unicode, guaranteeing the prompt text doesn't shift around as the drunk bishop walks.